Ai hopefully uses the split now + tile changes + AI tweaks
AI has been changed to hopefully use the split properly (both before and after). Can't properly test at the moment though.

Tiles now better track who is on them to allow the AI to make better decisions.

Misc AI tweaks for decision making when it comes to landing on a damage space with an enemy character. This required adding a function to the gamemanager class.